+++ CHANGELOG.rst
@@ -15,6 +15,11 @@ paths are considered internals and can change in minor and patch releases.
 v4.39.0 (2025-03-??)
 --------------------
 
+Added
+^^^^^
+- Resolve parameters completely from stubs when ``inspect.signature`` fails
+  (`#698 <https://github.com/omni-us/jsonargparse/pull/698>`__).
+
 Changed
 ^^^^^^^
 - Untyped parameters with ``None`` default no longer skipped when
@@ -29,8 +34,6 @@ Added
 ^^^^^
 - Support ``shtab`` completion of ``Literal`` types (`#693
   <https://github.com/omni-us/jsonargparse/pull/693>`__).
-- Support for parsing optionals as positionals (`#692
-  <https://github.com/omni-us/jsonargparse/pull/692>`__).
 
 Changed
 ^^^^^^^